Add quality rate script and create a new directory to set scripts
